# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fleet Control API
  version: 0.2.1
  description: |
    Public web API consumed by the Fleet web UI. The API aggregates multiple
    device roles (audio players, video/CEC, Zigbee, cameras) and exposes a
    consistent contract for UI integrations. Schemas align with the live device
    capabilities matrix to ensure client compatibility.
  contact:
    name: Fleet Platform Team
    email: platform@fleet.example
servers:
  - url: /api
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  headers:
    CorrelationId:
      description: Correlation identifier for tracing requests end-to-end.
      schema:
        type: string
    RateLimitLimit:
      description: Maximum number of requests allowed in the current window.
      schema:
        type: integer
        format: int32
    RateLimitRemaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
        format: int32
    RetryAfterSeconds:
      description: Seconds until clients may retry after a 429 response.
      schema:
        type: integer
        format: int32
...
  responses:
    UnauthorizedError:
      description: Authentication failed or credentials missing.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: UNAUTHENTICATED
                message: Authentication token missing or invalid.
                correlationId: 00000000-0000-0000-0000-000000000000
    ForbiddenError:
      description: Authenticated user does not have permission to access the resource.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Requested resource does not exist.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: One or more request parameters failed validation.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid:
              value:
                code: INVALID_REQUEST
                message: Provided payload failed validation.
                correlationId: 11111111-1111-1111-1111-111111111111
    RateLimitError:
      description: Request rate limit exceeded.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
        x-ratelimit-limit:
          $ref: '#/components/headers/RateLimitLimit'
        x-ratelimit-remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        retry-after:
          $ref: '#/components/headers/RetryAfterSeconds'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadGatewayError:
      description: Upstream device returned an invalid response or is unreachable.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    GatewayTimeoutError:
      description: Upstream device timed out while processing the request.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Unexpected server error occurred.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotImplementedError:
      description: Endpoint contract defined but backend implementation is pending.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - correlationId
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable summary of the error.
        hint:
          type: string
          nullable: true
          description: Optional client-facing recommendation.
        correlationId:
          type: string
          format: uuid
          description: Correlates the request across systems.
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Additional contextual information for debugging.
...
    ZigbeeDevice:
      type: object
      required:
        - id
        - name
        - type
        - state
        - lastSeen
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        state:
          type: string
          enum:
            - open
            - closed
            - active
            - inactive
        lastSeen:
          type: string
          format: date-time
        battery:
          type: integer
          format: int32
          nullable: true
    ZigbeePairingCandidate:
      type: object
      required:
        - id
        - model
        - manufacturer
        - signal
        - lastSeen
      properties:
        id:
          type: string
        model:
          type: string
        manufacturer:
          type: string
        signal:
          type: integer
          format: int32
        lastSeen:
          type: string
          format: date-time
    ZigbeePairingState:
      type: object
      required:
        - active
        - discovered
      properties:
        active:
          type: boolean
        startedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        discovered:
          type: array
          items:
            $ref: '#/components/schemas/ZigbeePairingCandidate'
        confirmed:
          type: array
          items:
            type: string
    ZigbeeState:
      type: object
      required:
        - devices
        - hub
        - pairing
        - rules
      properties:
        devices:
          type: array
          items:
            $ref: '#/components/schemas/ZigbeeDevice'
        hub:
          type: object
          required:
            - id
            - status
            - channel
            - lastHeartbeatAt
          properties:
            id:
              type: string
            status:
              type: string
            channel:
              type: integer
              format: int32
            lastHeartbeatAt:
              type: string
              format: date-time
        pairing:
          $ref: '#/components/schemas/ZigbeePairingState'
        rules:
          type: array
          items:
            type: object
            required:
              - id
              - name
              - enabled
              - createdAt
              - updatedAt
            properties:
              id:
                type: string
              name:
                type: string
              enabled:
                type: boolean
              createdAt:
                type: string
                format: date-time
              updatedAt:
                type: string
                format: date-time
...
    ZigbeePairingStartRequest:
      type: object
      properties:
        durationSeconds:
          type: integer
          format: int32
          minimum: 5
          maximum: 900
...
paths:
...
  /zigbee/overview:
    get:
      tags:
        - Zigbee
      summary: Retrieve Zigbee hub state, devices, and quick actions.
      description: Retrieve Zigbee hub state, devices, and quick actions.
      operationId: getZigbeeOverview
      responses:
        '200':
          description: Zigbee overview payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZigbeeState'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '501':
          $ref: '#/components/responses/NotImplementedError'
        '502':
          $ref: '#/components/responses/BadGatewayError'
        '504':
          $ref: '#/components/responses/GatewayTimeoutError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /zigbee/pairing:
    post:
      tags:
        - Zigbee
      summary: Start Zigbee pairing mode for discovering new devices.
      description: Start Zigbee pairing mode for discovering new devices.
      operationId: startZigbeePairing
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZigbeePairingStartRequest'
      responses:
        '200':
          description: Pairing window started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZigbeePairingState'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '501':
          $ref: '#/components/responses/NotImplementedError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Zigbee
      summary: Stop Zigbee pairing mode.
      description: Stop Zigbee pairing mode.
      operationId: stopZigbeePairing
      responses:
        '200':
          description: Pairing mode stopped.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZigbeePairingState'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '501':
          $ref: '#/components/responses/NotImplementedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /zigbee/pairing/discovered:
    get:
      tags:
        - Zigbee
      summary: Poll for newly discovered Zigbee devices during pairing mode.
      description: Poll for newly discovered Zigbee devices during pairing mode.
      operationId: pollZigbeeDiscovered
      responses:
        '200':
          description: Discovered devices snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZigbeePairingState'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '501':
          $ref: '#/components/responses/NotImplementedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /zigbee/pairing/{deviceId}:
    post:
      tags:
        - Zigbee
      summary: Confirm pairing of a discovered Zigbee device.
      description: Confirm pairing of a discovered Zigbee device.
      operationId: confirmZigbeePairing
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Pairing confirmation accepted.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '501':
          $ref: '#/components/responses/NotImplementedError'
        '500':
          $ref: '#/components/responses/InternalError'
...